+++ tests/testthat/test-check_output.R
@@ -1,8 +1,6 @@
 library(testthat)
 library(loadflux)
 library(dplyr)
-library(tsibble)
-library(fabletools)
 
 test_that("check output classes", {
   output_table <- hydro_events(
@@ -13,67 +11,24 @@ test_that("check output classes", {
   )
 
   output_plot <- output_table %>%
-    event_plot(
-      q = discharge,
-      datetime = time,
-      he = he
-    )
+    event_plot(q = discharge, datetime = time, he = he)
 
   output_ti <- djanturb %>%
-    hydro_events(
-      q = discharge,
-      datetime = time,
-      window = 21
-    ) %>%
+    hydro_events(q = discharge, datetime = time, window = 21) %>%
     dplyr::filter(he == 2) %>%
-    TI(
-      ssc = discharge,
-      datetime = time
-    )
+    TI(ssc = discharge, datetime = time)
 
   output_ahi <- djan %>%
-    hydro_events(
-      q = discharge,
-      datetime = time,
-      window = 21
-    ) %>%
+    hydro_events(q = discharge, datetime = time, window = 21) %>%
     dplyr::filter(he == 2) %>%
-    AHI(
-      q = discharge,
-      ssc = SS
-    )
+    AHI(q = discharge, ssc = SS)
 
   output_shi <- djan %>%
-    hydro_events(
-      q = discharge,
-      datetime = time,
-      window = 21
-    ) %>%
+    hydro_events(q = discharge, datetime = time, window = 21) %>%
     dplyr::filter(he == 2) %>%
-    SHI(
-      q = discharge,
-      ssc = SS
-    )
+    SHI(q = discharge, ssc = SS)
 
-  output_feat <- djan %>%
-    hydro_events(
-      q = discharge,
-      datetime = time,
-      window = 21
-    ) %>%
-    as_tsibble(
-      key = he,
-      index = time
-    ) %>%
-    features(
-      time,
-      feat_event
-    )
-
-  expect_s3_class(output_plot,
-                  c("dygraphs", "htmlwidget"))
-  expect_s3_class(output_feat,
-                  c("tbl_df", "tbl", "data.frame"))
+  expect_s3_class(output_plot, c("dygraphs", "htmlwidget"))
   expect_type(output_ti, "double")
   expect_type(output_shi, "double")
   expect_type(output_ahi, "double")
@@ -89,9 +44,7 @@ test_that("Non-numeric or missing inputs should error", {
   expect_error(SHI(NA))
   expect_error(AHI(NA))
   expect_error(TI(NA))
-  expect_error(event_plot(dataframe = "cat",
-                          q = "cat",
-                          datetime = "cat"))
+  expect_error(event_plot(dataframe = "cat", q = "cat", datetime = "cat"))
 
   expect_error(SHI(NA))
 })